File and grouping restructure of RPC code.
